+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Fix the URL for the "Creating a Workflow file" link in the README. ([#247])
 
 ## [1.0.2] - 2020-08-21
 
@@ -147,3 +147,4 @@ Versioning].
 [#207]: https://github.com/ericcornelissen/svgo-action/pull/207
 [#223]: https://github.com/ericcornelissen/svgo-action/pull/223
 [#239]: https://github.com/ericcornelissen/svgo-action/pull/239
+[#247]: https://github.com/ericcornelissen/svgo-action/pull/247